+++ docs/source/change-log.rst
@@ -17,6 +17,7 @@ year-month-day
 
 * Added :py:meth:`~ipumspy.api.extract.extract_to_dict()` method to enable easy exporting of extract objects to dictionary objects for eventual writing to .json for ease of extract sharing.
 * Added :py:meth:`~ipumspy.api.exceptions.IpumsExtractNotSubmitted` exception
+* Added ability to download Stata, SPSS, SAS, and R command files with data files :py:meth:`~ipumspy.api.IpumsApiClient.download_extract()`.
 
 
 0.1.0